5101 Make separate properties package
